requirements.txt
# RAYVEN Dependencies

# Core
python-dotenv==1.0.0
requests==2.31.0

# Data & Analysis
pandas==2.0.3
numpy==1.24.3

# Technical Indicators
ta==0.10.2

# Authentication
PyJWT==2.8.0
cryptography==41.0.0

# Async Support
aiohttp==3.9.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
